Drag item to other inventory bugfix
Implemented self.move_inventory for items to check if they have changed inventory. If it is true then the inventory function Move_Item_To_New_Slot will not check for collision but simply remove the active item, reset the         self.move_inventory to False and return. The problem before was that the items got stuck in a limbo and existing in both inventories